pyproject.toml
[project]
name = "hybrid-search"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitriy",email = "dmitriy.trochshenko@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pypdf2 = "^3.0.1"
sentence-transformers = "^3.1.1"
langchain = "^0.3.1"
langchain-openai = "^0.2.2"
elasticsearch = "^8.15.0"
redis = "^5.0.8"
psycopg2-binary = "^2.9.9"
pgvector = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
